Complete daily china_daily spider with dictionaries
The spider has been completed using dictionaries instead of items, items will be used in the next version. All temporary reference files have been removed. A test-data directory has been added within the main scraper directory for storing test data
